- Removed RouterFactory and updated aliases - Updated tests to reflect new factory change - Removed unnecessary ServiceListener check - Changed string literal to class check in RouterConfigTrait.php
