+++ test/Http/HttpRouterFactoryTest.php
@@ -6,10 +6,20 @@ namespace LaminasTest\Router\Http;
 
 use Laminas\Router\Http\HttpRouterFactory;
 use Laminas\Router\RoutePluginManager;
-use LaminasTest\Router\RouterFactoryTest as TestCase;
+use Laminas\ServiceManager\ServiceManager;
+use LaminasTest\Router\TestAsset;
+use PHPUnit\Framework\TestCase;
+use Psr\Container\ContainerExceptionInterface;
+use Psr\Container\NotFoundExceptionInterface;
+
+use function array_merge_recursive;
 
 final class HttpRouterFactoryTest extends TestCase
 {
+    private array $defaultServiceConfig;
+
+    private HttpRouterFactory $factory;
+
     public function setUp(): void
     {
         $this->defaultServiceConfig = [
@@ -19,5 +29,47 @@ final class HttpRouterFactoryTest extends TestCase
         ];
 
         $this->factory = new HttpRouterFactory();
+    }
+
+    /**
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     */
+    public function testFactoryCanCreateRouterBasedOnConfiguredName(): void
+    {
+        $config   = array_merge_recursive($this->defaultServiceConfig, [
+            'services' => [
+                'config' => [
+                    'router' => [
+                        'router_class' => TestAsset\Router::class,
+                    ],
+                ],
+            ],
+        ]);
+        $services = new ServiceManager($config);
+
+        $router = $this->factory->__invoke($services, 'router');
+        $this->assertInstanceOf(TestAsset\Router::class, $router);
+    }
+
+    /**
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     */
+    public function testFactoryCanCreateRouterWhenOnlyHttpRouterConfigPresent(): void
+    {
+        $config   = array_merge_recursive($this->defaultServiceConfig, [
+            'services' => [
+                'config' => [
+                    'router' => [
+                        'router_class' => TestAsset\Router::class,
+                    ],
+                ],
+            ],
+        ]);
+        $services = new ServiceManager($config);
+
+        $router = $this->factory->__invoke($services, 'router');
+        $this->assertInstanceOf(TestAsset\Router::class, $router);
     }
 }
